+++ frontend/src/features/public-form/components/FormPaymentPage/stripe/components/PaymentBlock.tsx
